refactor: move paste rules to core
