+++ packages/core/src/extensions/paste-rule/mark-paste-rule.spec.ts
@@ -1,12 +1,9 @@
-import { defineMarkSpec, union } from '@prosekit/core'
 import { describe, expect, it } from 'vitest'
 
-import { defineBold } from '../bold/index.ts'
-import { defineDoc } from '../doc/index.ts'
-import { defineParagraph } from '../paragraph/index.ts'
-import { pasteHTML, pasteText } from '../testing/clipboard.ts'
-import { setupTestFromExtension } from '../testing/index.ts'
-import { defineText } from '../text/index.ts'
+import { union } from '../../editor/union.ts'
+import { defineMarkSpec } from '../../extensions/mark-spec.ts'
+import { pasteHTML, pasteText } from '../../testing/clipboard.ts'
+import { defineBold, defineDoc, defineParagraph, defineText, setupTestFromExtension } from '../../testing/index.ts'
 
 import type { MarkPasteRuleOptions } from './mark-paste-rule.ts'
 import { defineMarkPasteRule } from './mark-paste-rule.ts'
