+++ packages/core/src/extensions/paste-rule/mark-paste-rule.ts
@@ -1,6 +1,8 @@
-import { getMarkType, type PlainExtension } from '@prosekit/core'
 import type { Attrs, MarkType, ProseMirrorNode } from '@prosekit/pm/model'
 import { Fragment, Slice } from '@prosekit/pm/model'
+
+import type { PlainExtension } from '../../types/extension.ts'
+import { getMarkType } from '../../utils/get-mark-type.ts'
 
 import { definePasteRule } from './paste-rule.ts'
 import { splitTextByRegex } from './split-text-by-regex.ts'
