+++ packages/core/src/extensions/paste-rule/split-text-by-regex.spec.ts
