+++ packages/extensions/src/paste-rule/index.ts
@@ -1,2 +1,8 @@
-export { defineMarkPasteRule, type MarkPasteRuleOptions } from './mark-paste-rule.ts'
-export { definePasteRule, type PasteRuleHandler, type PasteRuleHandlerOptions, type PasteRuleOptions } from './paste-rule.ts'
+export {
+  defineMarkPasteRule,
+  definePasteRule,
+  type MarkPasteRuleOptions,
+  type PasteRuleHandler,
+  type PasteRuleHandlerOptions,
+  type PasteRuleOptions,
+} from '@prosekit/core'
